fix: fix some case for add bg func
