+++ packages/vchart-extension/src/charts/sequence-scatter/constant.ts
@@ -1,0 +1,2 @@
+export const DATA_KEY = 'dataKey';
+export const BACKGROUND_KEY = 'background';
